make GitFetch public
In Mach, we have optional dependencies: if you'd like to build from source,
you specify `-Ddawn-from-source=true` or a Zig struct field, and then we need
to clone the Dawn and DirectXShaderCompiler source code.

It's nice to not clone these sources by default, though, since they're a bit
large and using a binary build is the default if you don't specify that option.

In this case, we need a way to fetch these sources ourselves, so we make GitFetch
public.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>